Actions PORO'ları toplayan bir konvansiyon uygula
Beklediğimiz düzen şöyle (actions dizininin app/services altından app
altına taşınacağını varsayıyoruz):

	app/actions/«ilişkili model yolu»/«eylem grup adı»/«eylemler»...

Örneğin

	app/actions/prospective_student/upsert.rb

dosyası aday öğrencinin güncellenmesi (insert/update) eylemine karşı
düşüyor.

Bu örnekte ise (yer ve isimler tartışılabilir)

	app/actions/user/verification/send.rb

dosyası kullanıcının (model yolu) doğrulanması (eylem grup adı)
eylemlerinden "doğrulama kodunun gönderilmesi" eylemine karşı düşüyor.

Bu düzenlemede tecrübe ettiğimiz bir sıkıntı eylemlerde (send.rb,
verify.rb) kullanılacak PORO ve yardımcıları nereye koyacağımız
konusuydu.  Başlangıçta yaptığım başarısız bir girişimden sonra şöyle
bir konvansiyonun uygun olacağını düşündüm.

	app/actions/«ilişkili model yolu»/«eylem grup adı».rb

Buna göre

	app/actions/user/verification.rb

dosyası ilgili eylem grubundaki tüm PORO ve helperları barındırıyor ve

	app/actions/user/verification/

dizini sadece eylemlerden oluşuyor (keşfedilebilirliği artıran bir
husus).

Bu düzenlemenin zaafı birden fazla PORO class ve yardımcı varsa (ki bu
örnekte var) hepsinin tek dosyada konsolide edilmesi.  Fakat duruma göre
bu bir zaaf değil avantaj olarak da değerlendirilebilir: kullanılan tüm
yapıtaşlarının tek bir dosyada incelenebilmesi.

Alternatifler:

- PORO ve helperlar konsolide edilmeden eylem grubu dizininde kendi
  dosyalarına konulabilirdi.  Örneğin phone nesnesi:

	app/actions/user/verification/phone.rb

  Bu düzenlemedeki sıkıntı bu şekilde ortaya çıkan irili ufaklı
  dosyaların gerçek eylem dosyalarının görünürlüğünü azaltması.

	app/actions/user/verification/phone.rb # bir PORO
	app/actions/user/verification/send.rb  # bir eylem

- PORO ve helperlar eylem grubu dizininde tek bir dosyada konsolide
  edilebilirdi.  Örneğin:

	app/actions/user/verification/support.rb # veya utils.rb

  Bu düzenlemedeki sıkıntılar ise (1) ilk alternatif kadar olmasa bile
  eylem görünürlüğünü bir parça azaltması, (2) support, utils gibi çok
  jenerik çağrışımlı isimlendirmeler kullanılması.

Son olarak...  Giderek büyüme eğilimi gösteren Rails uygulamalarında
okunurluk ve maintain edilebilirliği korumak/ölçeklemek gerçekten ciddi
bir sorun.  Amaca uygun PORO kullanımı ve buna benzer konvansiyonlar bu
mücadeledeki en büyük silahımız.  Maalesef Rails bu konuda (concerns
dışında) açık ve dokümante edilmiş konvansiyonlar sunmuyor.  Bu nedenle
bu konuda net ve herkes tarafından kullanılan bir "best practice" yok
(araştırmalarım sonucunda ulaştığım yargı), fakat esinlenilebilecek
güzel uygulamalar var.  Örneğin exercism.io'nun arkasındaki

	https://github.com/exercism/v3-website

projesi (app/commands dizini; bizdeki actions ile benzer).  Sonuç olarak
bu konu tartışmaya açık.  Farklı fikirleriniz varsa dinlemek isterim.
