+++ app/services/actions/user/verification/verify.rb
@@ -4,8 +4,8 @@ module Actions
   module User
     module Verification
       class Verify < Base
-        def run(code)
-          result.error unless phone.verify?(code)
+        def call(code)
+          result.accordingly { phone.verify?(code) }
         rescue Phone::MissingVerifyRecordError
           result.error(:base, I18n.t('.verification.without_any_record'))
         end
